backend/indexer_semantique/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Semantic embeddings
sentence-transformers==2.7.0
transformers==4.39.0
torch>=2.1.0
numpy==1.26.4
huggingface-hub>=0.21.0

# Vector search
faiss-cpu==1.7.4
# faiss-gpu==1.7.4  # Uncomment for GPU support

# Database
psycopg2-binary==2.9.9

# Message queue
pika==1.3.2

# Text processing
nltk==3.8.1
spacy==3.7.5

# Utilities
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
structlog==23.2.0
python-json-logger==2.0.7
python-dotenv==1.0.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Optional: For advanced vector operations
# pgvector==0.2.4  # If using PostgreSQL vector extension
# chromadb==0.4.18  # Alternative to FAISS